shell: don't segfault when pressing super+k when no surface is focused

+++ src/shell.c
@@ -3271,12 +3271,17 @@ static void
 force_kill_binding(struct wl_seat *seat, uint32_t time, uint32_t key,
 		   void *data)
 {
+	struct wl_surface *focus_surface;
 	struct wl_client *client;
 	pid_t pid;
 	uid_t uid;
 	gid_t gid;
 
-	client = seat->keyboard->focus->resource.client;
+	focus_surface = seat->keyboard->focus;
+	if (!focus_surface)
+		return;
+
+	client = focus_surface->resource.client;
 	wl_client_get_credentials(client, &pid, &uid, &gid);
 
 	kill(pid, SIGKILL);
